move *.go files to the v4/export directory
