+++ v4/export/mock.go
@@ -1,5 +1,3 @@
-//+build !codes
-
 package export
 
 import (
